Changes to source data members, started implementation on book inText
I added a new data member ArrayList<String[]> authorInfo:

authorInfo[a] carries info for authors[a]

authorInfo[0] is the surname
the rest is the first name, middle name, etc.

I also started implementing the inText() method for the book type. It is still WIP

One other thing is that I added comments to specify what each info string for each type contains to reference in the future